Gen/URL_shortener/backend/requirements.txt
Flask==3.1.1
Flask-CORS==6.0.0
SQLAlchemy==2.0.41
python-dotenv==1.1.0
shortuuid==1.0.13 